Reorganise kubernetes subpackage
